pyproject.toml
[tool.poetry]
name = "schedule-api"
version = "0.1.0"
description = ""
authors = ["Amer-On <a20051772@yandex.ru>"]

[tool.black]
line-length = 160
skip-string-normalization = true
include = '\.py$'
exclude = '''
(
    /(
        \.eggs
      | \.git
      | \.mypy_cache
      | deploy
      | \tests
      | \.venv
      | \sql
    )/
  | /(
        __pycache__
      | _mypy_cache
    )/
)'''


[tool.mypy]
ignore_missing_imports = true
disallow_untyped_calls = false
check_untyped_defs = true
warn_redundant_casts = false
show_error_codes = true
pretty = true
ignore_errors = false
exclude = 'tests'


[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
lines_after_imports = 2
lines_between_sections = 1
known_local_folder  = ['app']
known_first_party = ['fastapi', 'starlette', 'pydantic', 'pytest', 'uvicorn',
                      'httpx', 'tortoise', 'aiofile', 'bs4', 'apscheduler',
                      'redis', 'jose', 'schedule_parser']
sections = ['FUTURE', 'STDLIB', 'FIRSTPARTY', 'THIRDPARTY', 'LOCALFOLDER']

[tool.semgrep]
additional_args = "--severity=error"
paths = ["app/"]
exclude_paths = [".venv","./app/tests/*.py"]


[tool.poetry.dependencies]
python = "~3.12"
httpx = "^0.26.0"
fastapi = "^0.109.0"
pyyaml = "~6.0"
uvicorn = {extras = ["uvloop"], version = "^0.26.0"}
gunicorn = "~21.2.0"
aerich = {extras = ["asyncpg"], version = "^0.7.2"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
bcrypt = "^4.1.2"
passlib = "1.7.4"
aiofile = "^3.8.8"
lxml = "^5.1.0"
beautifulsoup4 = "^4.12.3"
redis = "^5.0.1"
python-jose = "^3.3.0"

[tool.poetry.group.test]
optional=true

[tool.poetry.group.test.dependencies]
pytest = "~7.4.2"
pytest-mock = "~3.11.1"
pytest-asyncio = "~0.21.1"

[tool.poetry.group.dev]
optional=true

[tool.poetry.group.dev.dependencies]
flake8 = "~6.1.0"
black = "~23.9.1"
pre-commit = "~3.4.0"
mkdocs = "~1.5.3"
mkdocs-material = "~9.4.4"
coverage = "~7.3.2"
flake8-quotes = "~3.3.2"
gitlint = "~0.19.1"
flake8-print = "~5.0.0"
isort = "~5.12.0"

[tool.poetry.group.scheduler]
optional=true

[tool.poetry.group.scheduler.dependencies]
python = "~3.12"
schedule-parser = {git = "git@github.com:PROF-Development/schedule-parser.git"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
apscheduler = "^3.10.4"
redis = "^5.0.1"
pyyaml = "~6.0"
beautifulsoup4 = "^4.12.3"
httpx = "^0.26.0"
aiofile = "^3.8.8"
bcrypt = "^4.1.2"
passlib = "1.7.4"
aerich = {extras = ["asyncpg"], version = "^0.7.2"}
aioredis = "^2.0.1"
lxml = "^5.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "app"

[tool.aerich]
tortoise_orm = "internal.repositories.db.helpers.TORTOISE_ORM_CONFIG"
location = "app/internal/repositories/db/migrations"
src_folder = "./app"
